fix: время обновления токена, класс ответа плейлистов

+++ app/src/main/java/com/wimank/pbfs/rest/response/RefreshTokenResponse.kt
@@ -10,7 +10,7 @@ data class RefreshTokenResponse(
     val accessToken: String?,
 
     @Json(name = "expires_in")
-    val expiresIn: Long?,
+    val expiresIn: Int?,
 
     @Json(name = "refresh_token")
     val refreshToken: String?,
